update paths providers and add new provider paypal and config initial
